Fix CaptureAll and support for RawM WithResource
We fix the implementation of the CaptureAll instance. Previously paths
would never be matched since the instance did not consume the rest of
the path like `CaptureAll` does. The rest of the path is now captured
and replaced with a `*` place holder and this is also the case for
enumerating the endpoint.

We also add instances for `RawM` and `WithResource` and add a test case
to the spec to check that `CaptureAll` and `RawM` behave as expected.